requirements.txt
Django==1.7.4
pinax-theme-bootstrap==5.7.1
django-user-accounts==1.0.1
metron==1.3.5
eventlog==0.9.0
django-jsonfield==0.9.13

# pinax-blog==3.1.3
pinax-lms-activities==0.2.0

gondor==1.2.5
gunicorn==19.1.1
psycopg2==2.5.4

greek-accentuation==0.8
